refactored version now working
Spreadsheet submissions weren't working as the array of values needed converting from an array of arrays to a single array. Gnarly! Thankfully we can use a prototype .concat function to get what we want.
This should allw now be much more usable, however I'm still not sure why it's not posting to google sheets sometimes.