Colleague assignment problems
Now resolved if colleague assignment is a one time thing, which it isn't. Colleague assignment depends on prevColleague values in shift state. Resetting it on assignment complete invalidates the assignment complete stage. Likely need to use the assign state with three values, but initialising in a sensible way is hard.