Mostly working POC
Can post answers fine but rendering answers at the end of the journey is failing due to async requests in a functional component.
May need a better way to save answers as current model creates an answer object on page load, and CP is always the first page, create many redundant answer objects!
Probably should continue with saving answers in state and then posting them all in a single request at the end of the journey