GraphQL now successfully posts CP Answers
The problem with my Data Model was GraphQL was set to be nested 2 levels deep, and my Data Model was going 3 levels deep. The other issue is AWS expect a GraphQL entity to post to DynamoDB regardless of whether you need to. So I used AWSJSON to save an object to the answers table, which for basic confirmation practice queries should be enough. I hope!